+++ highlighters/main/test-data/array-cmdsep3.zsh
@@ -32,8 +32,10 @@ BUFFER=$'a=( foo \n bar )'
 
 expected_region_highlight=(
   '1 3 assign' # a=(
+  '3 3 reserved-word' # (
   '5 7 default' # foo
   '9 9 commandseparator' # \n
   '11 13 default' # bar
   '15 15 assign' # )
+  '15 15 reserved-word' # )
 )