+++ highlighters/main/test-data/assign-array3.zsh
@@ -31,9 +31,11 @@ BUFFER='A=(hello world) b=42'
 
 expected_region_highlight=(
   "1 3 assign" # A=(
+  "3 3 reserved-word" # (
   "4 8 default" # hello
   "10 14 default" # world
   "15 15 assign" # )
+  "15 15 reserved-word" # )
   "17 20 assign" # b=42
   "19 20 default" # 42
 )